feat(processor): wip integration
